requirements.txt
imutils==0.5.4
matplotlib==3.3.4
numpy==1.19.5
opencv_python==4.6.0.66
Pillow==9.4.0
scikit_learn==1.2.0
tensorflow==2.6.2
torch==1.10.2
tqdm==4.64.1
